+++ packages/jotai/src/utility-types/SetAtom.ts
@@ -1,0 +1,1 @@
+export type SetAtom<TArgs extends unknown[], TResult> = (...args: TArgs) => TResult