+++ packages/jotai/src/utility-types/index.ts
@@ -1,1 +1,2 @@
 export type { ChildrenRenderProps } from './ChildrenRenderProps'
+export type { SetAtom } from './SetAtom'