+++ packages/cache/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @suspensive/cache
 
+## 0.3.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @suspensive/utils@2.13.0
+
+## 0.3.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @suspensive/utils@2.12.3
+
 ## 0.3.5
 
 ### Patch Changes