+++ packages/react-image/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @suspensive/react-image
 
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @suspensive/utils@2.13.0
+
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @suspensive/utils@2.12.3
+
 ## 0.0.26
 
 ### Patch Changes